requirements.txt
openai
python-dotenv
scikit-learn
google-generativeai
anthropic